Move init to pipeline
